+++ examples/magento-graphcms/components/Layout/LayoutNavigation.tsx
@@ -1,40 +1,11 @@
-import { CartFab, useCartEnabled } from '@graphcommerce/magento-cart'
-import { magentoMenuToNavigation } from '@graphcommerce/magento-category'
-import { CustomerFab, CustomerMenuFabItem } from '@graphcommerce/magento-customer'
-import { SearchFab, SearchField } from '@graphcommerce/magento-search'
-import { WishlistFab, WishlistMenuFabItem } from '@graphcommerce/magento-wishlist'
-import {
-  DesktopNavActions,
-  DesktopNavBar,
-  iconCustomerService,
-  iconHeart,
-  NavigationFab,
-  MenuFabSecondaryItem,
-  PlaceholderFab,
-  IconSvg,
-  DesktopNavItem,
-  DarkLightModeMenuSecondaryItem,
-  iconChevronDown,
-  NavigationProvider,
-  NavigationOverlay,
-  useNavigationSelection,
-  useMemoDeep,
-  MobileTopRight,
-} from '@graphcommerce/next-ui'
-import { t } from '@lingui/core/macro'
-import { Trans } from '@lingui/react/macro'
-import { Divider, Fab } from '@mui/material'
+import { CartFab } from '@graphcommerce/magento-cart'
+import { NavigationFab, useNavigationSelection } from '@graphcommerce/next-ui'
 import { useRouter } from 'next/router'
 import { Footer } from './Footer'
-import { LayoutQuery } from './Layout.gql'
-import { Logo } from './Logo'
-import { productListRenderer } from '../ProductListItems/productListRenderer'
-import {
-  StoreSwitcherButton,
-  StoreSwitcherFab,
-  StoreSwitcherMenuFabSecondaryItem,
-} from '@graphcommerce/magento-store'
+import { Header } from './Header'
+import type { LayoutQuery } from './Layout.gql'
 import { LayoutDefault, type LayoutDefaultProps } from './LayoutDefault'
+import { MenuOverlay } from './MenuOverlay'
 
 export type LayoutNavigationProps = LayoutQuery &
   Omit<LayoutDefaultProps, 'footer' | 'header' | 'cartFab' | 'menuFab'>
@@ -44,124 +15,15 @@ export function LayoutNavigation(props: LayoutNavigationProps) {
 
   const selection = useNavigationSelection()
   const router = useRouter()
-  const cartEnabled = useCartEnabled()
 
   return (
     <>
-      <NavigationProvider
-        selection={selection}
-        items={useMemoDeep(
-          () => [
-            { id: 'home', name: <Trans>Home</Trans>, href: '/' },
-            {
-              id: 'manual-item-one',
-              href: `/${menu?.items?.[0]?.children?.[0]?.url_path}`,
-              name: menu?.items?.[0]?.children?.[0]?.name ?? '',
-            },
-            {
-              id: 'manual-item-two',
-              href: `/${menu?.items?.[0]?.children?.[1]?.url_path}`,
-              name: menu?.items?.[0]?.children?.[1]?.name ?? '',
-            },
-            ...magentoMenuToNavigation(menu, true),
-            { id: 'blog', name: 'Blog', href: '/blog' },
-            <Divider key='divider' sx={(theme) => ({ my: theme.spacings.xs })} />,
-            <CustomerMenuFabItem
-              onClick={() => selection.set(false)}
-              key='account'
-              guestHref='/account/signin'
-              authHref='/account'
-            >
-              <Trans>Account</Trans>
-            </CustomerMenuFabItem>,
-            <MenuFabSecondaryItem
-              key='service'
-              icon={<IconSvg src={iconCustomerService} size='medium' />}
-              href='/service'
-            >
-              <Trans>Customer Service</Trans>
-            </MenuFabSecondaryItem>,
-            <WishlistMenuFabItem
-              onClick={() => selection.set(false)}
-              key='wishlist'
-              icon={<IconSvg src={iconHeart} size='medium' />}
-            >
-              <Trans>Wishlist</Trans>
-            </WishlistMenuFabItem>,
-            <DarkLightModeMenuSecondaryItem key='darkmode' />,
-            <StoreSwitcherMenuFabSecondaryItem key='store-switcher' />,
-          ],
-          [menu, selection],
-        )}
-      >
-        <NavigationOverlay
-          stretchColumns={false}
-          variantSm='left'
-          sizeSm='full'
-          justifySm='start'
-          itemWidthSm='70vw'
-          variantMd='left'
-          sizeMd='full'
-          justifyMd='start'
-          itemWidthMd='230px'
-          mouseEvent='hover'
-          itemPadding='md'
-        />
-      </NavigationProvider>
+      <MenuOverlay menu={menu} selection={selection} />
 
       <LayoutDefault
         {...uiProps}
         noSticky={router.asPath.split('?')[0] === '/'}
-        header={
-          <>
-            <Logo />
-
-            <DesktopNavBar>
-              {menu?.items?.[0]?.children?.slice(0, 2).map((item) => (
-                <DesktopNavItem key={item?.uid} href={`/${item?.url_path}`}>
-                  {item?.name}
-                </DesktopNavItem>
-              ))}
-
-              <DesktopNavItem
-                onClick={() => selection.set([menu?.items?.[0]?.uid || ''])}
-                onKeyUp={(evt) => {
-                  if (evt.key === 'Enter') {
-                    selection.set([menu?.items?.[0]?.uid || ''])
-                  }
-                }}
-                tabIndex={0}
-              >
-                {menu?.items?.[0]?.name}
-                <IconSvg src={iconChevronDown} />
-              </DesktopNavItem>
-
-              <DesktopNavItem href='/blog'>
-                <Trans>Blog</Trans>
-              </DesktopNavItem>
-            </DesktopNavBar>
-
-            <DesktopNavActions>
-              <SearchField
-                formControl={{ sx: { width: '400px' } }}
-                searchField={{ productListRenderer }}
-              />
-              <StoreSwitcherButton />
-              <Fab href='/service' aria-label={t`Customer Service`} size='large' color='inherit'>
-                <IconSvg src={iconCustomerService} size='large' />
-              </Fab>
-              <WishlistFab icon={<IconSvg src={iconHeart} size='large' />} />
-              <CustomerFab guestHref='/account/signin' authHref='/account' />
-              {/* The placeholder exists because the CartFab is sticky but we want to reserve the space for the <CartFab /> */}
-              {cartEnabled && <PlaceholderFab />}
-            </DesktopNavActions>
-
-            <MobileTopRight>
-              <StoreSwitcherFab size='responsiveMedium' />
-              <SearchFab size='responsiveMedium' />
-            </MobileTopRight>
-          </>
-        }
+        header={<Header menu={menu} selection={selection} />}
         footer={<Footer footer={footer} />}
         cartFab={<CartFab />}
         menuFab={<NavigationFab onClick={() => selection.set([])} />}
