+++ examples/magento-storyblok/components/Layout/LayoutMinimal.tsx
@@ -1,9 +1,9 @@
-import type { LayoutDefaultProps } from '@graphcommerce/next-ui'
-import { LayoutDefault } from '@graphcommerce/next-ui'
 import { GlobalConfigProvider } from '../Storyblok/GlobalConfigProvider'
 import type { StoryblokGlobalConfig } from '../Storyblok/types'
 import { Footer } from './Footer'
+import { HeaderContainer } from './HeaderContainer'
 import type { LayoutQuery } from './Layout.gql'
+import { LayoutDefault, type LayoutDefaultProps } from './LayoutDefault'
 import { Logo } from './Logo'
 
 export type LayoutMinimalProps = LayoutQuery &
@@ -18,7 +18,11 @@ export function LayoutMinimal(props: LayoutMinimalProps) {
     <GlobalConfigProvider value={globalConfig}>
       <LayoutDefault
         {...uiProps}
-        header={<Logo />}
+        header={
+          <HeaderContainer>
+            <Logo />
+          </HeaderContainer>
+        }
         footer={<Footer />}
         sx={(theme) => ({ background: theme.vars.palette.background.paper })}
       >
